Change: Move models into abstract.py
Closes #38
