+++ src/swerex/runtime/abstract.py
@@ -1,20 +1,99 @@
 from abc import ABC, abstractmethod
 
-from swerex.models import (
-    Action,
-    CloseSessionRequest,
-    Command,
-    CommandResponse,
-    CreateSessionRequest,
-    CreateSessionResponse,
-    Observation,
-    ReadFileRequest,
-    ReadFileResponse,
-    UploadRequest,
-    UploadResponse,
-    WriteFileRequest,
-    WriteFileResponse,
-)
+from pydantic import BaseModel
+
+
+class CreateSessionRequest(BaseModel):
+    session: str = "default"
+    # Source the following files before running commands.
+    # The reason this gets a special treatment is that these files
+    # often overwrite PS1, which we need to reset.
+    startup_source: list[str] = []
+
+
+class CreateSessionResponse(BaseModel):
+    success: bool = True
+    failure_reason: str = ""
+    output: str = ""
+
+
+# todo: implement non-output-timeout
+class Action(BaseModel):
+    command: str
+    session: str = "default"
+    timeout: float | None = None
+    # For a non-exiting command to an interactive program
+    # (e.g., gdb), set this to True.
+    # This will disable checking for exit codes, since the command won't terminate.
+    # If the command is something like "quit" and should terminate the
+    # interactive program, set this to False.
+    is_interactive_command: bool = False
+    is_interactive_quit: bool = False
+    # Outputs to expect in addition to the PS1
+    expect: list[str] = []
+
+
+class Observation(BaseModel):
+    success: bool = True
+    output: str = ""
+    exit_code: int | None = None
+    failure_reason: str = ""
+    # Which of the expect strings was matched to terminate the command.
+    # Empty string if the command timed out etc.
+    expect_string: str = ""
+
+
+class CloseSessionRequest(BaseModel):
+    session: str = "default"
+
+
+class CloseSessionResponse(BaseModel):
+    success: bool = True
+    failure_reason: str = ""
+
+
+class Command(BaseModel):
+    command: str | list[str]
+    timeout: float | None = None
+    shell: bool = False
+
+
+class CommandResponse(BaseModel):
+    stdout: str = ""
+    stderr: str = ""
+    exit_code: int | None = None
+    success: bool = True
+    failure_reason: str = ""
+
+
+class ReadFileRequest(BaseModel):
+    path: str
+
+
+class ReadFileResponse(BaseModel):
+    success: bool = True
+    failure_reason: str = ""
+    content: str = ""
+
+
+class WriteFileRequest(BaseModel):
+    content: str
+    path: str
+
+
+class WriteFileResponse(BaseModel):
+    success: bool = True
+    failure_reason: str = ""
+
+
+class UploadRequest(BaseModel):
+    source_path: str
+    target_path: str
+
+
+class UploadResponse(BaseModel):
+    success: bool = True
+    failure_reason: str = ""
 
 
 class AbstractRuntime(ABC):
